# LZ4 set acceleration parameter