--- a/unreleased_history/public_api_changes/compression_options_level_lz4.md
+++ b/unreleased_history/public_api_changes/compression_options_level_lz4.md
@@ -1,0 +1,1 @@
+* When using LZ4 compression, the `acceleration` parameter is configurable by setting the negated value in `CompressionOptions::level`. For example, `CompressionOptions::level=-10` will set `acceleration=10`